+++ src/browser/api_protocol/api_handlers/middleware_entity_existence.ts
@@ -4,6 +4,7 @@ import { appByUuid, windowExists } from '../../core_state';
 import { applicationApiMap } from './application.js';
 import { MessagePackage } from '../transport_strategy/api_transport_base';
 import { windowApiMap } from './window.js';
+import { webContentsApiMap } from './webcontents';
 
 const apisToIgnore = new Set([
     // Application
@@ -53,6 +54,14 @@ function verifyEntityExistence(msg: MessagePackage, next: () => void): void {
             return nack('Could not locate the requested application');
         }
 
+    } else if (webContentsApiMap.hasOwnProperty(action)) {
+         // Window API
+
+        const wndExists = windowExists(uuid, name);
+
+        if (!wndExists) {
+            return nack('Could not locate the requested contents');
+        }
     } else if (windowApiMap.hasOwnProperty(action)) {
         // Window API
