+++ src/browser/api_protocol/api_handlers/window.ts
@@ -1,5 +1,4 @@
 
-const { Application } = require('../../api/application');
 const { Window } = require('../../api/window');
 import {
     getGroupingWindowIdentity,
@@ -30,7 +29,6 @@ export const windowApiMap = {
     'disable-window-frame': disableUserMovement,
     'dock-window': dockWindow,
     'enable-window-frame': enableUserMovement,
-    'execute-javascript-in-window': { apiFunc: executeJavascript, apiPath: '.executeJavaScript' },
     'flash-window': flashWindow,
     'focus-window': focusWindow,
     'get-current-window-options': getCurrentWindowOptions,
@@ -42,7 +40,6 @@ export const windowApiMap = {
     'get-window-options': getWindowOptions,
     'get-window-snapshot': { apiFunc: getWindowSnapshot, apiPath: '.getSnapshot' },
     'get-window-state': getWindowState,
-    'get-zoom-level': getZoomLevel,
     'hide-window': hideWindow,
     'is-window-showing': isWindowShowing,
     'join-window-group': joinWindowGroup,
@@ -52,12 +49,7 @@ export const windowApiMap = {
     'minimize-window': minimizeWindow,
     'move-window': moveWindow,
     'move-window-by': moveWindowBy,
-    'navigate-window': navigateWindow,
-    'navigate-window-back': navigateWindowBack,
-    'navigate-window-forward': navigateWindowForward,
-    'stop-window-navigation': stopWindowNavigation,
     'register-window-name': registerWindowName,
-    'reload-window': reloadWindow,
     'redirect-window-to-url': redirectWindowToUrl, // Deprecated
     'resize-window': resizeWindow,
     'resize-window-by': resizeWindowBy,
@@ -67,7 +59,6 @@ export const windowApiMap = {
     'set-foreground-window': setForegroundWindow,
     'set-window-bounds': setWindowBounds,
     'set-window-preload-state': setWindowPreloadState,
-    'set-zoom-level': setZoomLevel,
     'show-at-window': showAtWindow,
     'stop-flash-window': stopFlashWindow,
     'undock-window': undockWindow,
@@ -222,51 +213,6 @@ function moveWindowBy(identity: Identity, message: APIMessage, ack: Acker): void
     const windowIdentity = getTargetWindowIdentity(payload);
 
     Window.moveBy(windowIdentity, deltaLeft, deltaTop);
-    ack(successAck);
-}
-
-function navigateWindow(identity: Identity, message: APIMessage, ack: Acker, nack: (error: Error) => void): void {
-    const { payload } = message;
-    const { url } = payload;
-    const windowIdentity = getTargetWindowIdentity(payload);
-
-    return Window.navigate(windowIdentity, url)
-        .then(() => ack(successAck))
-        .catch(nack);
-}
-
-function navigateWindowBack(identity: Identity, message: APIMessage, ack: Acker, nack: (error: Error) => void): void {
-    const { payload } = message;
-    const windowIdentity = getTargetWindowIdentity(payload);
-
-    Window.navigateBack(windowIdentity)
-        .then(() => ack(successAck))
-        .catch(nack);
-}
-
-function navigateWindowForward(identity: Identity, message: APIMessage, ack: Acker, nack: (error: Error) => void): void {
-    const { payload } = message;
-    const windowIdentity = getTargetWindowIdentity(payload);
-
-    Window.navigateForward(windowIdentity)
-        .then(() => ack(successAck))
-        .catch(nack);
-}
-
-function stopWindowNavigation(identity: Identity, message: APIMessage, ack: Acker): void {
-    const { payload } = message;
-    const windowIdentity = getTargetWindowIdentity(payload);
-
-    Window.stopNavigation(windowIdentity);
-    ack(successAck);
-}
-
-function reloadWindow(identity: Identity, message: APIMessage, ack: Acker): void {
-    const { payload } = message;
-    const { ignoreCache } = payload;
-    const windowIdentity = getTargetWindowIdentity(payload);
-
-    Window.reload(windowIdentity, ignoreCache);
     ack(successAck);
 }
 
@@ -435,32 +381,6 @@ function enableUserMovement(identity: Identity, message: APIMessage, ack: Acker)
     ack(successAck);
 }
 
-function executeJavascript(identity: Identity, message: APIMessage, ack: Acker, nack: (error: Error) => void): void {
-    const { payload } = message;
-    const { code } = payload;
-    const dataAck = Object.assign({}, successAck);
-    const windowIdentity = getTargetWindowIdentity(payload);
-    let { uuid: pUuid } = windowIdentity;
-
-    while (pUuid) {
-        if (pUuid === identity.uuid) {
-            return Window.executeJavascript(windowIdentity, code, (err: Error, result: any) => {
-                if (err) {
-                    nack(err); // TODO: this nack doesn't follow the protocol
-                } else {
-                    dataAck.data = result;
-                    ack(dataAck);
-                }
-            });
-        }
-        pUuid = Application.getParentApplication({
-            uuid: pUuid
-        });
-    }
-
-    return nack(new Error('Rejected, target window is not owned by requesting identity'));
-}
-
 function disableUserMovement(identity: Identity, message: APIMessage, ack: Acker): void {
     const { payload } = message;
     const windowIdentity = getTargetWindowIdentity(payload);
@@ -537,26 +457,6 @@ function getCachedBounds(identity: Identity, message: APIMessage, ack: Acker, na
         dataAck.data = data;
         ack(dataAck);
     }, nack);
-}
-
-function getZoomLevel(identity: Identity, message: APIMessage, ack: Acker): void {
-    const { payload } = message;
-    const dataAck = Object.assign({}, successAck);
-    const windowIdentity = getTargetWindowIdentity(payload);
-
-    Window.getZoomLevel(windowIdentity, (result: number) => {
-        dataAck.data = result;
-        ack(dataAck);
-    });
-}
-
-function setZoomLevel(identity: Identity, message: APIMessage, ack: Acker): void {
-    const { payload } = message;
-    const { level } = payload;
-    const windowIdentity = getTargetWindowIdentity(payload);
-
-    Window.setZoomLevel(windowIdentity, level);
-    ack(successAck);
 }
 
 function registerWindowName(identity: Identity, message: APIMessage, ack: Acker): void {
