+++ src/browser/api_protocol/index.ts
@@ -25,6 +25,7 @@ import {
     registerMiddleware as registerProcessExternalAppMiddleware,
     legacyWindowingEnabled
 } from './api_handlers/deprecated_external_windowing_middleware';
+import { init as initWebContentsHandler } from './api_handlers/webcontents';
 
 
 // Middleware registration. The order is important.
@@ -53,8 +54,8 @@ export function initApiHandlers() {
     initSystemApiHandler();
     const globalHotkeyApiHandler = new GlobalHotkeyApiHandler();
     initWindowApiHandler();
+    initWebContentsHandler();
     initExternalWindowApiHandler();
-
     initApiProtocol();
 
     const apiPolicyProcessor = require('./api_handlers/api_policy_processor');
